Provide tracelib and private scripts
These are some pickle scripts created for new feature investigations.  They're
very specialized, and not likely of tremendous use to anyone else, which is
why they're starting off in a private branch here.

.pylintrc: I'm using pylint for script work.
tracelib.py: A library that encapsulates the boilerplate that most
   pickle scripts will need.  This might eventually be promoted to the
   main branch.  All the pickle scripts will use it.  It derived from
   unpickle.py.
scan.py: the same as unpickle.py - a script that calculates function
   use frequencies - but implemented using tracelib.
blobs.py: looks for coalescable data blobs.
disorder.py: looks for out-of-order commands.