apitrace: fix a premature termination bug in the trim operation
The trim operation has a feature where once the highest needed command
number is passed, the operation will terminate.  Unfortunately, multi-threaded
applications typically produce traces with a small number of out-of-order
commands, e.g. instead of commands [0, 1, 2], the trace could have commands
in the order [0, 2, 1].  (This is because command numbers are assigned upon
function entry, but commands are written to the trace file on function exit; if
one function is entered on on thread, and another enters and exits on another
thread before the first one finishes, commands will be out-of-order.)

This means the optimization can cause the trim operation to terminate too
early if commands are out of order.  E.g., for the trace [0, 2, 1] listed
above, if requested to trim "0-1", the resulting trace will actually only have
command 0 inside it, because the operation will terminate upon seeing command 2.

This fix tracks out-of-order command numbers, and ensures that the trim operation
continues until all appropriate command numbers have been seen and processed,
even if they're out-of-order.