First implementation of delta debugging
This is the start of an investigation into using the delta debugging algorithm
for minimizing traces.  We're still finding issues, so it's not complete yet.

DD.py: came from Andreas Zeller.  No changes have been made except to import it.

mintrace.py: a standalone Python script for invoking DD with apitrace.