Fixed trim bugs found during review, and repaired trim control flow
Jose noted that the algorithm to keep track of the last seen call number
breaks down if a call is skipped in the middle; changed the control
flow so the call number is always registered, even if it's about
to be skipped.

Jose also noted that the "goto" was not necessary.  Repaired the control
flow.

There's a small change in behavior: in the earlier code, the frame number
was not updated if a frame-updating call was skipped because it wasn't
on the right thread.  I think this is an oversight - if frame numbers aren't
calculated consistently, then the user-specified frame numbers could be off.
I changed the behavior to always update the frame number, which I think is more
sane.